Issue #14: moved validate_clerk_input function out of checkout.py to a new file called validate_clerk_input.py
